requirements.txt
streamlit==1.42.0
Pillow==11.1.0
reportlab==3.6.9
